backend/requirements.txt
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
MarkupSafe==1.1.1
nltk==3.5
numpy==1.20.1
pandas==1.2.3
pdfminer==20191125
pdfminer.six==20201018
psycopg2-binary==2.8.6
pycparser==2.20
pycryptodome==3.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
regex==2020.11.13
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sortedcontainers==2.3.0
threadpoolctl==2.1.0
tqdm==4.58.0
Werkzeug==1.0.1
